third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-invalid-args-expected.txt

Ensure correct behavior of createImageBitmap for invalid inputs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS createImageBitmap(image) threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS createImageBitmap(video) threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS createImageBitmap(undefined) threw exception TypeError: Type error.
PASS createImageBitmap(image, 0, 0, 10, 0) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS createImageBitmap(image, 0, 0, 0, 10) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS createImageBitmap(video, 0, 0, 10, 0) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS createImageBitmap(video, 0, 0, 0, 10) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS createImageBitmap(aCanvas, 0, 0, 10, 0) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS createImageBitmap(aCanvas, 0, 0, 0, 10) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS createImageBitmap(d, 0, 0, 10, 0) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS createImageBitmap(d, 0, 0, 0, 10) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS createImageBitmap(aCtx, 0, 0, 10, 0) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS createImageBitmap(aCtx, 0, 0, 0, 10) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS createImageBitmap(testBitmap, 0, 0, 10, 0) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS createImageBitmap(testBitmap, 0, 0, 0, 10) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS createImageBitmap(blob, 0, 0, 10, 0) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS createImageBitmap(blob, 0, 0, 0, 10) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS createImageBitmap(invalidBlob) was rejected.
PASS successfullyParsed is true

TEST COMPLETE
